requirements.txt
Django==1.7.6
django-crispy-forms==1.4.0
django-grappelli==2.6.3
djangorestframework==3.1.0
gnureadline==6.3.3
ipython==3.0.0
psycopg2==2.6
wsgiref==0.1.2
